+++ user/application/error.go
@@ -1,4 +1,4 @@
-package event
+package application
 
 import "fmt"
