--- user/middleware/middleware.go
@@ -1,9 +1,0 @@
-package middleware
-
-import (
-	"github.com/WeisswurstSystems/WWM-BB/user"
-)
-
-type MiddlewareHandler struct {
-	UserStore user.Store
-}
